requirements.txt
chromadb>=0.5.0
streamlit==1.28.1
pydantic==2.5.0
python-dotenv==1.0.0
pypdf==4.0.1
ollama
